ActionPlan12 9-12-23

+++ Plutone_Project/src/main/java/main/VeerificaTesto.java
@@ -68,7 +68,7 @@ tot++;
 
         
         
-        return tot/(daverificare.size()+listavero.size());
+        return ((double)(tot))/((double)(daverificare.size()+listavero.size()));
 
     }
